pyproject.toml
[project]
name = "runeform"
version = "0.1.0"
description = "Compositional layout engine for brand marketing"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "pydantic>=2.0",
    "pillow>=11.0",
    "anthropic>=0.52",
    "python-multipart>=0.0.20",
    "beautifulsoup4>=4.12",
    "httpx>=0.28",
    "python-dotenv>=1.2.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
